# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "API for the Laravel E-Commerce application. Provides endpoints for browsing products, managing carts/wishlists, placing orders, and customer account management.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Application server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List the authenticated user orders",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{orderNumber}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show a single order by order number",
                "operationId": "7960d10ba66849383b294d5132175474",
                "parameters": [
                    {
                        "name": "orderNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Checkout: convert cart to an order",
                "operationId": "c90fce9fa5e9860be3ce8ede472afc4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipping_address",
                                    "payment_method"
                                ],
                                "properties": {
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "123 Main St, City, Country",
                                        "maxLength": 500
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "cash_on_delivery",
                                        "enum": [
                                            "cash_on_delivery",
                                            "bank_transfer",
                                            "credit_card"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order placed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order placed successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Empty cart or insufficient stock"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Wireless Headphones"
                    },
                    "slug": {
                        "type": "string",
                        "example": "wireless-headphones"
                    },
                    "image_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 79.99
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 159.98
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-ABC123XYZ"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "confirmed",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "paid",
                            "failed",
                            "refunded"
                        ]
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "cash_on_delivery",
                        "enum": [
                            "cash_on_delivery",
                            "bank_transfer",
                            "credit_card"
                        ]
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 159.98
                    },
                    "shipping_fee": {
                        "type": "number",
                        "format": "float",
                        "example": 5
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 164.98
                    },
                    "shipping_address": {
                        "type": "string",
                        "example": "123 Main St, City"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items_count": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "OrderDetail": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Order"
                    },
                    {
                        "properties": {
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OrderItem"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 16
                    },
                    "total": {
                        "type": "integer",
                        "example": 80
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token from POST /api/login or POST /api/register",
                "scheme": "bearer"
            }
        }
    },
...
}